+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [TBD] - TBD
 ### Changed
-* Upgraded `cascading` to 3.3.0 (was 3.0.2).
+- Upgraded `cascading` to 3.3.0 (was 3.0.2).
+- Use a copy of the `Tuple` in `TapDataReader` rather than the original.
 
 ## [3.0.2] - 2015-09-18
 ### Added